+++ IdleGame/Assets/Resources/Data/SummonTable.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 12ea1f939d85b82408d0e5627b067fe4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
