+++ IdleGame/Assets/Resources/Data/SummonTable/EquipmentSummonTable.json.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: fa84cf996a5199640bf3cd93a3b19fc4
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
