+++ IdleGame/Assets/Resources/ExcelTable.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7c12ff5ca2d7edb4f982fdad27982b7e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
