+++ IdleGame/Assets/Resources/ExcelTable/SummonTable.xlsx.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c920f27bcdf8caa4bbd2533e80de44c9
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
