+++ IdleGame/Assets/Resources/ExcelTable/~$SummonTable.xlsx.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 248acfe8d3532b443b88bb78cb3284d4
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
