+++ IdleGame/Assets/Scripts/Main/Model/Shop/SummonManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4315dd0628ff5264f9f25887ef3f0b55
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
