+++ IdleGame/Assets/Scripts/Managers/Manager.cs
@@ -39,6 +39,7 @@ public class Manager : MonoBehaviour
     private readonly StageManager stage = new();
     private readonly InventoryManager inventory = new();
     private readonly QuestManager quest = new();
+    private readonly SummonManager summon = new();
 
     public static ResourceManager Resource => Instance != null ? Instance.resource : null;
     public static DataManager Data => Instance != null ? Instance.data : null;
@@ -48,5 +49,6 @@ public class Manager : MonoBehaviour
     public static StageManager Stage => Instance != null ? Instance.stage : null;
     public static InventoryManager Inventory => Instance != null ? Instance.inventory : null;
     public static QuestManager Quest => Instance != null ? Instance.quest : null;
+    public static SummonManager Summon => Instance != null ? Instance.summon : null;
     #endregion
 }
